clu-ser/requirements.txt
blinker==1.9.0
bson==0.5.10
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
Flask==3.1.0
Flask-Cors==5.0.0
fonttools==4.56.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.0
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
pymongo==4.11
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
threadpoolctl==3.5.0
tzdata==2025.1
Werkzeug==3.1.3
